Added camera temperature to the OSD
With the new mavfwd of Tipo Man we can now have OSD temperature of the camera.
The -t option will inject the temperature of the camera to the MAVLINK IMU messages and then we will be able to display it on the OSD.
The -p 100 will enable a 3-way switch to be used for changing modes remotely. 
For 6-way switch increase the delay value and setup the transmitter accordingly.